Еще немного правок

+++ src/main/java/com/kyd3snik/travel/model/Resort.java
@@ -27,7 +27,7 @@ public class Resort {
     private List<Tag> tags;
     @ManyToOne(fetch = FetchType.LAZY)
     private Hotel hotel;
-    private int durationInDays;
+    private long durationInDays;
     private Date startDate;
     private Date endDate;
     private float cost;
